+++ src/pages/index.tsx
@@ -1,8 +1,7 @@
 //[TODO]インデックス丸ごといらないのだが。。。。
 import React, { useContext } from "react";
 
-import SignInComponent from "./sign-in";
-import SignUpComponent from "./sign-up";
+import SignInComponent from "./auth";
 import Dashboard from "./dashboard";
 import { AuthContext } from "../utils/authProvider";
