+++ src/pages/sign-up.tsx
@@ -1,164 +1,164 @@
 import React, { useState } from "react";
-import { useDispatch } from "react-redux";
-import { useRouter } from "next/router";
-import Avatar from "@material-ui/core/Avatar";
-import Button from "@material-ui/core/Button";
-import CssBaseline from "@material-ui/core/CssBaseline";
-import TextField from "@material-ui/core/TextField";
-import FormControlLabel from "@material-ui/core/FormControlLabel";
-import Checkbox from "@material-ui/core/Checkbox";
-import Link from "@material-ui/core/Link";
-import Grid from "@material-ui/core/Grid";
-import Box from "@material-ui/core/Box";
-import LockOutlinedIcon from "@material-ui/icons/LockOutlined";
-import Typography from "@material-ui/core/Typography";
-import { makeStyles } from "@material-ui/core/styles";
-import Container from "@material-ui/core/Container";
+// import { useDispatch } from "react-redux";
+// import { useRouter } from "next/router";
+// import Avatar from "@material-ui/core/Avatar";
+// import Button from "@material-ui/core/Button";
+// import CssBaseline from "@material-ui/core/CssBaseline";
+// import TextField from "@material-ui/core/TextField";
+// import FormControlLabel from "@material-ui/core/FormControlLabel";
+// import Checkbox from "@material-ui/core/Checkbox";
+// import Link from "@material-ui/core/Link";
+// import Grid from "@material-ui/core/Grid";
+// import Box from "@material-ui/core/Box";
+// import LockOutlinedIcon from "@material-ui/icons/LockOutlined";
+// import Typography from "@material-ui/core/Typography";
+// import { makeStyles } from "@material-ui/core/styles";
+// import Container from "@material-ui/core/Container";
 
-import { setCurrentUser, setSignInUp } from "../utils/reducers";
-import { auth } from "../../firebaseClient";
+// import { setCurrentUser, setSignInUp } from "../utils/reducers";
+// import { auth } from "../../firebaseClient";
 
-function Copyright() {
-  return (
-    <Typography variant="body2" color="textSecondary" align="center">
-      {"Copyright © "}
-      <Link color="inherit" href="https://material-ui.com/">
-        House-moving Manager
-      </Link>{" "}
-      {new Date().getFullYear()}
-      {"."}
-    </Typography>
-  );
-}
+// function Copyright() {
+//   return (
+//     <Typography variant="body2" color="textSecondary" align="center">
+//       {"Copyright © "}
+//       <Link color="inherit" href="https://material-ui.com/">
+//         House-moving Manager
+//       </Link>{" "}
+//       {new Date().getFullYear()}
+//       {"."}
+//     </Typography>
+//   );
+// }
 
-const useStyles = makeStyles((theme) => ({
-  paper: {
-    marginTop: theme.spacing(8),
-    display: "flex",
-    flexDirection: "column",
-    alignItems: "center",
-  },
-  avatar: {
-    margin: theme.spacing(1),
-    backgroundColor: theme.palette.secondary.main,
-  },
-  form: {
-    width: "100%",
-    marginTop: theme.spacing(1),
-  },
-  submit: {
-    margin: theme.spacing(3, 0, 2),
-  },
-}));
+// const useStyles = makeStyles((theme) => ({
+//   paper: {
+//     marginTop: theme.spacing(8),
+//     display: "flex",
+//     flexDirection: "column",
+//     alignItems: "center",
+//   },
+//   avatar: {
+//     margin: theme.spacing(1),
+//     backgroundColor: theme.palette.secondary.main,
+//   },
+//   form: {
+//     width: "100%",
+//     marginTop: theme.spacing(1),
+//   },
+//   submit: {
+//     margin: theme.spacing(3, 0, 2),
+//   },
+// }));
 
-export default function SignUpComponent() {
-  const classes = useStyles();
-  const router = useRouter();
-  const dispatch = useDispatch();
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [userName, setUserName] = useState("");
+// export default function SignUpComponent() {
+//   // const classes = useStyles();
+//   // const router = useRouter();
+//   // const dispatch = useDispatch();
+//   // const [email, setEmail] = useState("");
+//   // const [password, setPassword] = useState("");
+//   // const [userName, setUserName] = useState("");
 
-  const SignUp = async (e: any) => {
-    e.preventDefault();
-    try {
-      await auth
-        .createUserWithEmailAndPassword(email, password)
-        .then((user) => {
-          user.user && dispatch(setCurrentUser(user.user.uid));
-        })
-        .then(() => router.push("/dashboard"))
-        .catch((err) => console.error(err));
-    } catch (err) {
-      alert(err.message);
-    }
-  };
+//   // const SignUp = async (e: any) => {
+//   //   e.preventDefault();
+//   //   try {
+//   //     await auth
+//   //       .createUserWithEmailAndPassword(email, password)
+//   //       .then((user) => {
+//   //         user.user && dispatch(setCurrentUser(user.user.uid));
+//   //       })
+//   //       .then(() => router.push("/dashboard"))
+//   //       .catch((err) => console.error(err));
+//   //   } catch (err) {
+//   //     alert(err.message);
+//   //   }
+//   // };
 
-  return (
-    <Container component="main" maxWidth="xs">
-      <CssBaseline />
-      <div className={classes.paper}>
-        <Avatar className={classes.avatar}>
-          <LockOutlinedIcon />
-        </Avatar>
-        <Typography component="h1" variant="h5">
-          ユーザー登録
-        </Typography>
-        <form className={classes.form} noValidate onSubmit={SignUp}>
-          <TextField
-            variant="outlined"
-            margin="normal"
-            fullWidth
-            name="username"
-            label="ユーザー名"
-            id="username"
-            autoComplete="username"
-            autoFocus
-            value={userName}
-            onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
-              setUserName(e.target.value);
-            }}
-          />
+//   return (
+//     <Container component="main" maxWidth="xs">
+//       <CssBaseline />
+//       <div className={classes.paper}>
+//         <Avatar className={classes.avatar}>
+//           <LockOutlinedIcon />
+//         </Avatar>
+//         {/* <Typography component="h1" variant="h5">
+//           ユーザー登録
+//         </Typography>
+//         <form className={classes.form} noValidate onSubmit={SignUp}>
+//           <TextField
+//             variant="outlined"
+//             margin="normal"
+//             fullWidth
+//             name="username"
+//             label="ユーザー名"
+//             id="username"
+//             autoComplete="username"
+//             autoFocus
+//             value={userName}
+//             onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
+//               setUserName(e.target.value);
+//             }}
+//           />
 
-          <TextField
-            variant="outlined"
-            margin="normal"
-            required
-            fullWidth
-            id="email"
-            label="メールアドレス"
-            name="email"
-            autoComplete="email"
-            autoFocus
-            value={email}
-            onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
-              setEmail(e.target.value);
-            }}
-          />
-          <TextField
-            variant="outlined"
-            margin="normal"
-            required
-            fullWidth
-            name="password"
-            label="パスワード"
-            type="password"
-            id="password"
-            autoComplete="current-password"
-            value={password}
-            onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
-              setPassword(e.target.value);
-            }}
-          />
-          <FormControlLabel
-            control={<Checkbox value="remember" color="primary" />}
-            label="ユーザー情報を記憶しますか？"
-          />
-          <Button
-            type="submit"
-            fullWidth
-            variant="contained"
-            color="primary"
-            className={classes.submit}
-          >
-            登録
-          </Button>
-          <Grid container>
-            <Grid item xs></Grid>
-            <Grid item>
-              <Link
-                onClick={() => dispatch(setSignInUp("signIn"))}
-                href="/sign-in"
-              >
-                ログインはこちら
-              </Link>
-            </Grid>
-          </Grid>
-        </form>
-      </div>
-      <Box mt={8}>
-        <Copyright />
-      </Box>
-    </Container>
-  );
-}
+//           <TextField
+//             variant="outlined"
+//             margin="normal"
+//             required
+//             fullWidth
+//             id="email"
+//             label="メールアドレス"
+//             name="email"
+//             autoComplete="email"
+//             autoFocus
+//             value={email}
+//             onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
+//               setEmail(e.target.value);
+//             }}
+//           />
+//           <TextField
+//             variant="outlined"
+//             margin="normal"
+//             required
+//             fullWidth
+//             name="password"
+//             label="パスワード"
+//             type="password"
+//             id="password"
+//             autoComplete="current-password"
+//             value={password}
+//             onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
+//               setPassword(e.target.value);
+//             }}
+//           />
+//           <FormControlLabel
+//             control={<Checkbox value="remember" color="primary" />}
+//             label="ユーザー情報を記憶しますか？"
+//           />
+//           <Button
+//             type="submit"
+//             fullWidth
+//             variant="contained"
+//             color="primary"
+//             className={classes.submit}
+//           >
+//             登録
+//           </Button>
+//           <Grid container>
+//             <Grid item xs></Grid>
+//             <Grid item>
+//               <Link
+//                 onClick={() => dispatch(setSignInUp("signIn"))}
+//                 href="/sign-in"
+//               >
+//                 ログインはこちら
+//               </Link>
+//             </Grid>
+//           </Grid>
+//         </form> */}
+//       </div>
+//       <Box mt={8}>
+//         <Copyright />
+//       </Box>
+//     </Container>
+//   );
+// }
